add srp for pages/CourseList; move magic numbers to constants folder
