+++ src/constants/Pagination.js
@@ -1,0 +1,3 @@
+const COURSES_PER_PAGE = 10;
+
+export { COURSES_PER_PAGE };
